Set build number to git revisions + 1000.
I've made a proper mess of this. I didn't realise that CFBundleVersion must be
an integer, so when I set it to the Git revision, I think Sparkle has
interpreted it as an integer. In the case of 1.3.3, the last released version,
the revisions is 423e2d4, which is interpreted as 423. We need to make sure
that later versions are higher than this, so I'm just adding 1000 to add the
build numbers. Sorry.